requirements.txt
bcrypt==4.3.0

fastapi==0.115.12

passlib==1.7.4

python-jose==3.5.0

SQLAlchemy==2.0.41

uvicorn==0.34.3
